requirements.txt
# Core ML / DL
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0

# YOLO
ultralytics>=8.0.0

# DETR (HuggingFace Transformers)
transformers>=4.35.0

# Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Numerical / Science
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Graph utilities
networkx>=3.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
plotly>=5.17.0

# Data management
PyYAML>=6.0
tqdm>=4.65.0
pandas>=2.0.0

# Experiment tracking (optional)
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
